hw/nvme: add boot partition support
NVMe Boot Partitions provides an area that may be read by the host
without initializing queues or even enabling the controller. This allows
various platform initialization code to be stored on the NVMe device
instead of some separete medium.

This patch adds the read support for such an area, as well as support
for updating the boot partition contents from the host through the FW
Download and Commit commands.

Signed-off-by: Gollu Appalanaidu <anaidu.gollu@samsung.com>
[k.jensen: rebased]
Signed-off-by: Klaus Jensen <k.jensen@samsung.com>