hw/nvme: refactor/cleanup
Apart from some random small fixes here and there, the major thing here
is cleaning up how we handle command sets. Prior to this series, the
controller would not correctly validate namespace command sets against
CC.CSS. This is fixed here.

The most clean way of doing this (as far as I could tell) was to make
sure an nvme-subsys device exists (creating it if necessary). This
allows us to "store" the namespaces in the subsystem, using existing
functionality, and attach supported namespaces when the device is
started (instead of when the device is created/realized).

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: Keith Busch <kbusch@kernel.org>
To: Klaus Jensen <its@irrelevant.dk>
To: Jesper Devantier <foss@defmacro.it>
To: qemu-block@nongnu.org
To: qemu-devel@nongnu.org
Signed-off-by: Klaus Jensen <k.jensen@samsung.com>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20241216-nvme-queue-v1-0-4e42212b92f7@samsung.com



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20241216-nvme-queue-f4151c5d7507",
    "prefixes": [],
    "history": {
      "v1": [
        "20241216-nvme-queue-v1-0-4e42212b92f7@samsung.com"
      ]
    }
  }
}